+++ src/api.cpp
@@ -9,7 +9,7 @@
 
 using namespace webcachesim;
 
-Interface::Interface(const string &cache_type, const int &cache_size, const map<string, string> &params) {
+Interface::Interface(const string &cache_type, const uint64_t &cache_size, const map<string, string> &params) {
     string webcachesim_cache_type;
     if (cache_type == "LRU") {
         webcachesim_cache_type = "ParallelLRU";
